# Flatten declarative record instance creation lookups